Changed ADDON_VIEWS_COUNT from a constant to a property so that the value can be changed as needed in the application.
Edge case: if there are a lot of items of largely varying sizes, one of the items which should be off-screen when it is being recycled from top-to-bottom or vice versa flickers back and forth between both states. Adding more objects here helps hide this off screen. In my test case, I increased this to 16, but the exact number will vary based on the particular data.